pyproject.toml
[project]
name = "ok"
version = "0.1.0"
description = "project to fetch research papers"
authors = [
    {name = "Vipin Kumar",email = "vipinkhushuk@email.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "typer (>=0.15.1,<0.16.0)",
    "biopython (>=1.84,<2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
